backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
rdflib==7.0.0
pyshacl==0.26.0
httpx==0.27.2
SPARQLWrapper==2.0.0
email-validator==2.2.0
redis==5.0.8
